autoClose Contribution
Added option for autoClose on html but set to false.

Added polyfills for hideEl and showEl at end of js file.

Added logic to hide claendar when date is set for both single date and date range on line 153 of js.

A placeholder for the element needs to be added so that the calendar can be accessed after the calendar is closed.